+++ builder_pkgs/build_web_compilers/lib/src/build_frontend_server/frontend_server_resources.dart
@@ -6,6 +6,7 @@ import 'dart:convert';
 
 import 'package:build/build.dart';
 import 'package:path/path.dart' as p;
+import 'package:scratch_space/scratch_space.dart';
 
 /// A persistent shared [FrontendServerState] for DDC workers that interact with
 /// the Frontend Server.
@@ -17,6 +18,15 @@ class FrontendServerState {
   /// This must be set before any asset builders run when compiling with DDC and
   /// hot reload.
   AssetId? entrypointAssetId;
+
+  /// The scratch space where the Frontend Server writes its outputs (`.js`,
+  /// `.map`, and `.metadata` files).
+  ///
+  /// These files are read by downstream builders to create build assets.
+  ///
+  /// When not null, the scratch space should be initialized over an existing
+  /// directory (rather than a fresh one).
+  ScratchSpace? fesScratchSpace;
 
   /// Looks for and loads a `.web.entrypoint.json` file if it exists.
   ///