+++ builder_pkgs/build_web_compilers/test/build_frontend_server/frontend_server_driver_test.dart
@@ -6,8 +6,8 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:build_web_compilers/src/build_modules/common.dart';
-import 'package:build_web_compilers/src/build_modules/frontend_server_driver.dart';
+import 'package:build_web_compilers/src/build_frontend_server/frontend_server_driver.dart';
+import 'package:build_web_compilers/src/common.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 
@@ -26,6 +26,10 @@ void main() {
         ..writeAsStringSync(
           jsonEncode({'configVersion': 2, 'packages': <String>[]}),
         );
+      final configFile = File(
+        p.join(Directory.current.path, fesManagerConfigPath),
+      );
+      if (configFile.existsSync()) configFile.deleteSync();
       server = await PersistentFrontendServer.start(
         sdkRoot: sdkDir,
         fileSystemRoot: tempDir.uri,
@@ -108,6 +112,10 @@ void main() {
         ..writeAsStringSync(
           jsonEncode({'configVersion': 2, 'packages': <String>[]}),
         );
+      final configFile = File(
+        p.join(Directory.current.path, fesManagerConfigPath),
+      );
+      if (configFile.existsSync()) configFile.deleteSync();
       server = await PersistentFrontendServer.start(
         sdkRoot: sdkDir,
         fileSystemRoot: tempDir.uri,
@@ -168,9 +176,7 @@ void main() {
       expect(output, isNotNull);
       expect(output!.errorCount, 0);
 
-      final jsOutputFile = File(
-        jsFESOutputPath.replaceFirst('.dart.lib.js', '.ddc.js'),
-      );
+      final jsOutputFile = File(jsFESOutputPath);
       expect(jsOutputFile.existsSync(), isTrue);
       final content = jsOutputFile.readAsStringSync();
       expect(content, contains('function main()'));